Configuration Manager working
