+++ ThirdPartyKeys.example.plist
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>ServiceKey</key>
+	<dict>
+		<key>Release</key>
+		<string></string>
+		<key>Staging</key>
+		<string></string>
+		<key>Debug</key>
+		<string></string>
+	</dict>
+</dict>
+</plist>
