+++ swift-ui-base/Managers/SessionManager.swift
